Naveen/approach 2 buffreader (#1)
* use BufReader

* make some changes

* Appraoch 2: Use BufferedRead and remove print

The print!() was in the hot path i.e it was in the for loop.
The print!() is not advised to be had in the hot path as it needs
to talk to OS to keep asking for locks. This slows down the process.

The better solution might have been to lock the stdin and then use it.
I did not try that. TODO: Keep that for future.

Time: ~20 minutes, 3 (218 s)  minutes in release mode [reconfirm]

* add notes