Use memchr and loop instead of standard iterator
Instead of using split function to split the lines, we can use
memchr. The split function does not use SIMD which is bad. The
memchr crate uses SIMD which fastens the splitting.

The split function is slow because, it apparently contrusct the
split parser everytime it is called. memchr is better than that.

TODO: Add the details of the blog and the comparision between split
and memchr from the github repo.

Time: 30s